continued reorganizing auth components, added scss, changed signupform and loginform styling
