+++ src/components/AuthBar/AuthBar.module.scss
@@ -25,7 +25,7 @@
   padding: 15px 32px;
   text-align: center;
   text-decoration: none;
-  border-radius: 8px;
+  border-radius: 10px;
 }
 
 .logIn-content {
@@ -39,5 +39,10 @@
   padding: 15px 28px;
   text-align: center;
   text-decoration: none;
-  border-radius: 8px;
+  border-radius: 10px;
+}
+
+.popup-content{
+  background-color: hotpink;
+  padding: 0 0;
 }
