+++ src/components/AuthGreet/AuthGreet.module.scss
@@ -1,0 +1,5 @@
+.authGreetContainer{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
