+++ src/components/LogInForm/LogInForm.module.scss
@@ -1,0 +1,5 @@
+.logInFormContainer {
+  display: flex;
+  justify-content: center;
+}
+
