+++ src/components/LoginPage/LoginPage.js
@@ -1,10 +1,11 @@
-import AuthGreet from '../../components/AuthGreet/AuthGreet'
-import LoginForm from '../../components/LogInForm/LogInForm'
+import AuthGreet from '../AuthGreet/AuthGreet'
+import LoginForm from '../LogInForm/LogInForm'
 // import { Routes, Route } from "react-router-dom"
+import styles from "../LoginPage/LoginPage.module.scss"
 
 export default function LoginPage(props){
     return(
-        <main>
+        <main className={styles.loginPageContainer}>
             <AuthGreet />
             <LoginForm setUser={props.setUser}></LoginForm>
             Sign in with Apple<br/>
