+++ src/components/LoginPage/LoginPage.module.scss
@@ -1,0 +1,3 @@
+// main {
+//   background-color: hotpink;
+// }
