+++ src/components/SignUpForm/SignUpForm.js
@@ -39,15 +39,19 @@ export default class SignUpForm extends Component {
         const disable = this.state.password !== this.state.confirm;
         return (
           <div>
-            <div className="form-container">
-              <form autoComplete="off" onSubmit={this.handleSubmit}>
+            <div className={styles.signUpFormContainer}>
+              <form 
+              autoComplete="off" 
+              onSubmit={this.handleSubmit}
+              >
+                <section className={styles.signUpSection}>
                 <label>First Name: </label>
-                <input type="text" name="firstName" value={this.state.firstName} onChange={this.handleChange} /><br/>
+                  <input type="text" name="firstName" value={this.state.firstName} onChange={this.handleChange} /><br/>
                 <label>Last Name: </label> 
-                <input type="text" name="lastName" value={this.state.lastName} onChange={this.handleChange} /><br/>
+                  <input type="text" name="lastName" value={this.state.lastName} onChange={this.handleChange} /><br/>
                 <label>Email: <span className={styles.required}>*</span></label>
-                <input type="email" name="email" value={this.state.email} onChange={this.handleChange} required /><br/>
-                <label>Username: <span className={styles.required}>*</span></label>
+                  <input type="email" name="email" value={this.state.email} onChange={this.handleChange} required /><br/>
+                <label >Username: <span className={styles.required}>*</span></label>
                 <input type="text" name="username" value={this.state.username} onChange={this.handleChange} required /><br/>
                 <label>Profile Photo: </label>
                 <input type="text" name="profilePhoto" value={this.state.profilePhoto} onChange={this.handleChange} /><br/>
@@ -55,7 +59,10 @@ export default class SignUpForm extends Component {
                 <input type="password" name="password" value={this.state.password} onChange={this.handleChange} required /><br/>
                 <label>Confirm: </label>
                 <input type="password" name="confirm" value={this.state.confirm} onChange={this.handleChange} required />
-                <button type="submit" disabled={disable}>SIGN UP</button>
+                </section>
+                <div className={styles.btnSection}>
+                  <button type="submit" disabled={disable}>SIGN UP</button>
+                </div>
               </form>
             </div>
             <p className="error-message">&nbsp;{this.state.error}</p>
