+++ src/components/SignupPage/SignupPage.js
@@ -1,5 +1,5 @@
-import AuthGreet from "../../components/AuthGreet/AuthGreet"
-import SignUpForm from "../../components/SignUpForm/SignUpForm"
+import AuthGreet from "../AuthGreet/AuthGreet"
+import SignUpForm from "../SignUpForm/SignUpForm"
 // import { Routes, Route } from "react-router-dom"
 
 export default function SignUpPage(props){
